Rewrite edit profiler to use a simpler algorithm w/exponential decay. Rename server file.
